requirements.txt
numpy==1.26.4
pillow==11.0.0
PyOpenGL==3.1.7
PyOpenGL-accelerate==3.1.7
PySDL2==0.9.16
